requirements.txt
Pillow==9.2.0
selenium==4.4.3